backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
rich
prompt_toolkit
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googlemaps
langchain
langchain-openai
langchain-community
chromadb
langchain-chroma
PyYAML
requests
